requirements.txt
torch
numpy
matplotlib
PyYAML
tqdm
pytest
